--- aibolit/ast/__init__.py
@@ -1,2 +1,0 @@
-from aibolit.ast.ast import AST  # noqa: F401
-from aibolit.ast.ast_node_type import ASTNodeType  # noqa: F401
