+++ aibolit/ast_framework/__init__.py
@@ -1,0 +1,2 @@
+from aibolit.ast_framework.ast import AST  # noqa: F401
+from aibolit.ast_framework.ast_node_type import ASTNodeType  # noqa: F401
