+++ aibolit/ast_framework/java_class_method.py
@@ -25,11 +25,11 @@ from typing import Dict, Set, TYPE_CHECKING
 from networkx import DiGraph, dfs_tree  # type: ignore
 
 from aibolit.utils.cfg_builder import build_cfg
-from aibolit.ast import AST, ASTNodeType
-from aibolit.ast.java_class_field import JavaClassField
+from aibolit.ast_framework import AST, ASTNodeType
+from aibolit.ast_framework.java_class_field import JavaClassField
 
 if TYPE_CHECKING:
-    from aibolit.ast.java_class import JavaClass
+    from aibolit.ast_framework.java_class import JavaClass
 
 
 class JavaClassMethod(AST):
