+++ aibolit/metrics/RFC/rfc.py
@@ -1,5 +1,5 @@
-from aibolit.ast import AST, ASTNodeType
-from aibolit.ast.java_package import JavaPackage
+from aibolit.ast_framework import AST, ASTNodeType
+from aibolit.ast_framework.java_package import JavaPackage
 from typing import Set, Any
