+++ aibolit/metrics/ncss/ncss.py
@@ -23,7 +23,7 @@
 from networkx import dfs_labeled_edges  # type: ignore
 
 from aibolit.utils.ast_builder import build_ast
-from aibolit.ast import AST, ASTNodeType
+from aibolit.ast_framework import AST, ASTNodeType
 
 
 class NCSSMetric():
