+++ aibolit/patterns/assert_in_code/assert_in_code.py
@@ -1,7 +1,7 @@
 from typing import List
 
 from aibolit.utils.ast_builder import build_ast
-from aibolit.ast import AST, ASTNodeType
+from aibolit.ast_framework import AST, ASTNodeType
 
 
 class AssertInCode:
