+++ aibolit/patterns/force_type_casting_finder/force_type_casting_finder.py
@@ -23,7 +23,7 @@
 from typing import List
 
 from aibolit.utils.ast_builder import build_ast
-from aibolit.ast import AST, ASTNodeType
+from aibolit.ast_framework import AST, ASTNodeType
 
 
 class ForceTypeCastingFinder:
