+++ aibolit/patterns/joined_validation/joined_validation.py
@@ -23,7 +23,7 @@
 from typing import List, Tuple
 
 from aibolit.utils.ast_builder import build_ast
-from aibolit.ast import AST, ASTNodeType
+from aibolit.ast_framework import AST, ASTNodeType
 
 
 class JoinedValidation:
