+++ aibolit/patterns/multiple_while/multiple_while.py
@@ -1,5 +1,5 @@
 from aibolit.utils.ast_builder import build_ast
-from aibolit.ast import AST, ASTNodeType
+from aibolit.ast_framework import AST, ASTNodeType
 
 
 class MultipleWhile:
