+++ aibolit/patterns/return_null/return_null.py
@@ -23,7 +23,7 @@
 from typing import List
 
 from aibolit.utils.ast_builder import build_ast
-from aibolit.ast import AST, ASTNodeType
+from aibolit.ast_framework import AST, ASTNodeType
 
 
 class ReturnNull:
