+++ aibolit/patterns/var_middle/var_middle.py
@@ -25,7 +25,7 @@ from networkx import DiGraph, dfs_labeled_edges  # type: ignore
 
 from aibolit.types_decl import LineNumber
 from aibolit.utils.ast_builder import build_ast
-from aibolit.ast import AST, ASTNodeType
+from aibolit.ast_framework import AST, ASTNodeType
 from aibolit.utils.scope_status import ScopeStatus, ScopeStatusFlags
