+++ test/utils/test_java_class.py
@@ -23,8 +23,8 @@
 from unittest import TestCase
 from pathlib import Path
 
-from aibolit.ast.java_package import JavaPackage
-from aibolit.ast import ASTNodeType
+from aibolit.ast_framework.java_package import JavaPackage
+from aibolit.ast_framework import ASTNodeType
 
 
 class JavaClassTestCase(TestCase):
