+++ test/utils/test_java_class_decomposition.py
@@ -23,8 +23,8 @@
 from unittest import TestCase
 from pathlib import Path
 
-from aibolit.ast.java_class_decomposition import decompose_java_class
-from aibolit.ast.java_package import JavaPackage
+from aibolit.ast_framework.java_class_decomposition import decompose_java_class
+from aibolit.ast_framework.java_package import JavaPackage
 
 
 class JavaClassDecompositionTestSuite(TestCase):
