+++ test/utils/test_java_package.py
@@ -25,8 +25,8 @@ from pathlib import Path
 
 from typing import Set
 
-from aibolit.ast import ASTNodeType
-from aibolit.ast.java_package import JavaPackage
+from aibolit.ast_framework import ASTNodeType
+from aibolit.ast_framework.java_package import JavaPackage
 
 
 class JavaPackageTestCase(TestCase):
